Move UserSettings and KnownModules to their own files
